refactor: remove "util" from filename of a few files in util directory
