+++ src/__tests__/commands/apps/oauth/update.test.ts
@@ -53,7 +53,7 @@ jest.unstable_mockModule('../../../../lib/command/util/apps-util.js', () => ({
 
 const oauthAppScopeDefMock = buildInputDefMock<string>('Scopes Mock')
 const redirectUrisDefMock = buildInputDefMock<string>('Redirect URIs Mock')
-jest.unstable_mockModule('../../../../lib/command/util/apps-util-input-primitives.js', () => ({
+jest.unstable_mockModule('../../../../lib/command/util/apps-input-primitives.js', () => ({
 	oauthAppScopeDef: oauthAppScopeDefMock,
 	redirectUrisDef: redirectUrisDefMock,
 }))
