+++ src/__tests__/commands/apps/update.test.ts
@@ -60,7 +60,7 @@ jest.unstable_mockModule('../../../lib/command/util/apps-user-input-update.js', 
 	getAppUpdateRequestFromUser: getAppUpdateRequestFromUserMock,
 }))
 
-jest.unstable_mockModule('../../../lib/command/util/apps-util-input-primitives.js', () => ({
+jest.unstable_mockModule('../../../lib/command/util/apps-input-primitives.js', () => ({
 	smartAppHelpText: 'smartapp help text',
 }))
